pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "locnus"
authors = [
    {name = "Jochen Wersdörfer", email = "jochen@wersdoerfer.de"},
]
dynamic = ["version", "description"]
requires-python = ">=3.11"
keywords = [
    "Django",
    "Mastodon",
    "ActivityPub",
    "Fediverse",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Topic :: Internet",
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "crispy-bootstrap5",
    "django",
    "django-htmx",
    "Mastodon.py",
    "whitenoise",
]

[project.optional-dependencies]
test = [
    "pytest >= 6",
    "pytest-cov >= 3",
    "pytest-django",
    "pytest-mock",
    "pytest-randomly",
]
doc = [
    "furo",
]
dev = [
    "black",
    "flake8",
    "django_extensions",
    "django-stubs",
    "isort",
    "jupyterlab",
    "jupyter-server<2.0.0",
    "rich",
    "tox",
    "typer",
]

[project.urls]
Documentation = "https://github.com/ephes/pytest-course"
Source = "https://github.com/ephes/pytest-course"

[tool.flit.module]
name = "locnus"

[tool.isort]
profile = "black"

[tool.black]
line-length = 119

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
minversion = "6.0"
testpaths = ["tests"]
filterwarnings = [
    "ignore::django.utils.deprecation.RemovedInDjango50Warning",
    "ignore::DeprecationWarning",
]
addopts = [
    "-ra",
    "--reuse-db",
    "--no-migrations",
    "--randomly-seed=1234",
]

[tool.coverage.run]
branch = true
source = ["locnus"]
omit = [
    "locnus/migrations/*",
    "locnus/settings.py",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
]
